Some small protocol fixes. Correct serialization of version and address messages.
Also require the height of the best chain to be specified when setting up a NetworkConnection. This API is getting too complicated and will be simplified soon. 

For now these fixes resolve a bug whereby the peer would not send us new blocks discovered during a session (as it thought we already had them).

git-svn-id: http://bitcoinj.googlecode.com/svn/trunk@58 a743e126-8abc-4207-e589-8f9b47be99bf